@@ -36,7 +36,11 @@
     "@types/express": "^4.17.14",
     "@types/node": "^18.11.9",
     "@types/node-fetch": "^2.6.2",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
+    "eslint": "^8.27.0",
     "ts-node": "^10.9.1",
-    "typescript": "^4.8.4"
+    "tsd": "^0.24.1",
+    "typescript": "^4.9.3"
   }
 }
